pyproject.toml
[tool.poetry]
name = "AIVUS-CAA"
version = "0.1.1"
description = ""
authors = ["yungselm", "cardionaut", "Pooya Mohammadi Kazaj"]
# package-mode = false

[tool.poetry.dependencies]
python = "3.10"
loguru = "*"
tqdm = "*"
openpyxl = "*"
matplotlib = "*"
seaborn = "*"
pylint = "*"
hydra-core = "*"
omegaconf = "*"
numpy = "1.26.4"
tensorflow = "2.15.0"
protobuf = ">=3.20.3"
monai = "*"
scikit-image = "*"
pydicom = "2.4.4"
customtkinter = "*"
pylibjpeg = "*"
pylibjpeg-libjpeg = "*"
pyqt5 = "5.15.2"
pyqtdarktheme = "2.1.0"
shapely = "*"
SimpleITK = "*"
opencv-python-headless = "4.9.0.80"
torch = "2.4.0"
torchvision = "0.19.0"
nnunetv2 = {git = "https://github.com/AI-in-Cardiovascular-Medicine/nnUZoo", rev = "main"}

[tool.poetry.group.dev.dependencies]
autohooks = "^23.1.0"
autohooks-plugin-black = "^22.11.0"
autohooks-plugin-isort = "^22.8.0"
pytest = "^7.2.1"
black = "^22.6.0"
isort = "^5.10.1"

[tool.autohooks]
mode = "poetry"
pre-commit = ["autohooks.plugins.black", "autohooks.plugins.isort"]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
skip-string-normalization = true

[tool.isort]
profile = "black"

[[tool.poetry.source]]
name = "pytorch_cpu"
url = "https://download.pytorch.org/whl/cpu"
priority = "explicit"